added api rest

+++ src/sockets.js
@@ -1,28 +1,5 @@
 import User from "./models/User";
-
-const getUserByPassword = async (password) => {
-  try {
-    const user = await User.findOne({ password });
-
-    if (!user) return null;
-
-    return {
-      position: user.position,
-      municipalityNumber: user.municipalityNumber,
-      lastName: user.lastName,
-      firstName: user.firstName,
-      gender: user.gender,
-      party: user.party,
-      startDate: user.startDate,
-      endDate: user.endDate,
-      memberStatus: user.memberStatus,
-      memberPhoto: user.memberPhoto,
-    };
-  } catch (error) {
-    console.error("Error al buscar usuario:", error);
-    return null;
-  }
-};
+import { getUserByPassword } from "./functions/get_user_by_password.js";
 
 export default (io) => {
   io.on("connection", (socket) => {
@@ -79,12 +56,10 @@ export default (io) => {
       const isPasswordUsed = users.some(
         (user) => user.password == userData.password
       );
-
       // Verificar si ya existe otro administrador
       const isAdminExists = users.some(
         (user) => user.position === "Administrador"
       );
-
       if (isPasswordUsed) {
         console.log("No se agrego el usuario, contraseña repetida");
         socket.emit("server:addusererror", "La contraseña ya está en uso.");
@@ -95,7 +70,7 @@ export default (io) => {
         User.create(userData);
         console.log("Usuario agregado");
         io.emit("server:adduser", userData);
-        socket.broadcast.emit("server:useradded", userData); // Emite el evento a todos los clientes para informar que se agregó un usuario
+        io.emit("server:getusers", users);
       }
     });
 
@@ -113,12 +88,17 @@ export default (io) => {
         console.log(password);
 
         if (!user) {
-          console.log("no borrado");
+          console.log("No se identifico ningun usuario");
         } else {
           // Eliminar el usuario
-          await User.deleteMany({ password: password });
-          console.log("Usuario borrado");
-
+          if (user.position === "Administrador") {
+            console.log("No puedes borrar un administrador");
+            io.emit("server:deleteerror", "No puedes borrar un administrador");
+            throw error;
+          } else {
+            await User.deleteMany({ password: password });
+            io.emit("server:deleteuser", user);
+          }
           const users = await User.find();
 
           io.emit("server:getusers", users); // Vuelve a emitir la lista de usuarios actualizada a todos los clientes
@@ -133,42 +113,3 @@ export default (io) => {
     });
   });
 };
-
-//LEGACY (No en uso)
-
-////(Dentro de default (io))
-// const emitNotes = async () => {
-//   const notes = await Note.find();
-//   io.emit("server:loadnotes", notes);
-// };
-// emitNotes();
-
-//import Note from "./models/Note";
-
-// //Cuando se cree una nueva nota
-// socket.on("client:newnote", async (data) => {
-//   const newNote = new Note(data);
-//   const savedNote = await newNote.save();
-//   console.log(savedNote);
-//   io.emit("server:newnote", savedNote);
-// });
-
-// socket.on("client:deletenote", async (id) => {
-//   await Note.findByIdAndDelete(id);
-//   console.log("Deleted: " + id);
-//   emitNotes();
-// });
-
-// socket.on("client:getnote", async (id) => {
-//   const note = await Note.findById(id);
-//   io.emit("server:selectednote", note);
-//   emitNotes();
-// });
-
-// socket.on("client:updatenote", async (updatedNote) => {
-//   await Note.findByIdAndUpdate(updatedNote._id, {
-//     title: updatedNote.title,
-//     description: updatedNote.description,
-//   });
-//   emitNotes();
-// });
